requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
boto3==1.18.29
botocore==1.21.29
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.6
django-bootstrap-form==3.4
django-filter==2.4.0
django-import-export==2.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docopt==0.6.2
et-xmlfile==1.1.0
fastapi==0.68.1
filelock==3.0.12
gTTS==2.2.3
h11==0.12.0
idna==3.2
jmespath==0.10.0
Js2Py==0.71
MarkupPy==1.14
MouseInfo==0.1.3
mysql-connector==2.2.9
mysqlclient @ file:///D:/Documents/Sampingan/Bisnis/%282021%29%20Flou%20Tools/development_v0.1/mysqlclient-2.0.3-cp39-cp39-win_amd64.whl
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
Pillow==8.3.1
pipwin==0.5.1
platformdirs==2.2.0
playsound==1.3.0
PyAudio @ file:///C:/Users/ASUS/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.53
pydantic==1.8.2
PyGetWindow==0.0.9
Pygments==2.10.0
pyjokes==0.6.0
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.27
pySmartDL==1.3.4
python-dateutil==2.8.2
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==5.1
pywin32==301
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
starlette==0.14.2
tablib==3.0.0
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
uvicorn==0.15.0
virtualenv==20.7.2
wikipedia==1.4.0
winspeech==1.0.1
xlrd==2.0.1
xlwt==1.3.0
